+++ src/telegram.rs
@@ -1,3 +1,4 @@
+pub(crate) mod albums;
 pub mod auth;
 pub mod client;
 pub mod converters;